+++ start/src/App.vue
@@ -1,15 +1,12 @@
 <template>
-  <f-checkbox-group v-model="option2" :change="change">
-    <f-checkbox label="鸡肉卷" />
-    <f-checkbox label="鸡排" />
-    <f-checkbox label="汉堡" />
-  </f-checkbox-group>
+  <f-switch v-model="value1" :change="change" />
+  <f-switch v-model="value2" />
 </template>
 
 <script lang="ts" setup>
   import { ref } from 'vue'
-
-  const option2 = ref(['汉堡'])
+  const value1 = ref(true)
+  const value2 = ref(false)
 
   function change(e) {
     console.log(e)